+++ tox.ini
@@ -7,25 +7,25 @@ deps =
     -rrequirements.dev.txt
 changedir = {envtmpdir}
 commands =
-    {envpython} -c "import topology_lib_sflow; print(topology_lib_sflow.__file__)"
+    {envpython} -c "import topology_lib_sflowtool; print(topology_lib_sflowtool.__file__)"
     flake8 {toxinidir}
     py.test \
         {posargs} \
         {toxinidir}/test \
-        {envsitepackagesdir}/topology_lib_sflow
+        {envsitepackagesdir}/topology_lib_sflowtool
 
 [testenv:coverage]
 basepython = python3.4
 commands =
     py.test \
         --junitxml=tests.xml \
-        --cov=topology_lib_sflow \
+        --cov=topology_lib_sflowtool \
         --cov-report xml \
         --cov-report html \
         --cov-report term \
         {posargs} \
         {toxinidir}/test \
-        {envsitepackagesdir}/topology_lib_sflow
+        {envsitepackagesdir}/topology_lib_sflowtool
 
 [testenv:doc]
 basepython = python3.4
